requirements.txt
bcrypt==3.2.0
cffi==1.14.6
click==7.1.2
Flask==1.1.2
gunicorn==20.1.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2-binary==2.9.1
pycparser==2.20
six==1.16.0
Werkzeug==1.0.1
